+++ HISTORY.md
@@ -1,3 +1,9 @@
+unreleased
+==========
+
+  * feat: add `contentTypeNegotiation` option which enables `text/plain` and `text/html`
+    responses based on the `Accept` header
+
 v2.1.1. / 2025-12-01
 ==================